new commit

+++ 14node_REST_APIs_Practice_Application/backend/controllers/feed.js
@@ -5,6 +5,7 @@ const { title } = require("process");
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -51,20 +52,30 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   // console.log("########", req.file.path);
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Maximilian" },
+    creator: req.userId,
   });
 
   post
     .save()
+    .then(result=>{
+      return User.findById(req.userId);
+    })
+    .then(user=>{
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then((result) => {
       // console.log(result);
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: {_id: creator._id, name: creator.name}
       });
     })
     .catch((err) => {
@@ -129,6 +140,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId){
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -163,13 +181,27 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if(post.creator.toString() !== req.userId){
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+
       // check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "Deleted post." });
+    .then(result=>{
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      console.log(user);
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result=>{
+      res.status(200).json({ message: "Deleted post." });      
     })
     .catch((err) => {
       if (!err.statusCode) {
